requirements.txt
backports.functools-lru-cache==1.5
cycler==0.10.0
decorator==4.3.0
dj-database-url==0.4.2
Django==1.11.13
django-bootstrap3==10.0.1
django-extensions==2.0.7
gunicorn==19.8.1
kiwisolver==1.0.1
matplotlib==2.2.2
networkx==2.1
numpy==1.14.3
opencv-python==3.4.1.15
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyparsing==2.2.0
pytesseract==0.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.1.0
six==1.11.0
subprocess32==3.5.1
typing==3.6.4
